+++ Objects/cooldown.h
@@ -12,13 +12,13 @@ class Cooldown
 {
 public:
     Cooldown(uint32_t id, uint32_t addr);
-    uint8_t id();
+    uint16_t id();
     CooldownGroup group();
     uint64_t endTime();
     bool isUp();
 
 private:
-    uint8_t m_id;
+    uint16_t m_id;
     uint32_t m_address;
 };
 }
